feat(cortex): S2 — consume metafactory-discord bundle; remove in-repo Discord CLI (#1173)
Formalizes the cortex-side decoupling from the Discord CLI tooling, which was
extracted to the-metafactory/metafactory-discord in S1 (ADR-0017, epic #1171).

Removed:
- src/cli/discord/ entirely (discord.ts + lib/ + skill/ + __tests__/).

Dependency declared:
- arc-manifest.yaml dependencies[]: metafactory-discord (>=0.1.0), with a note
  that registry auto-resolution is pending (repo-first per ADR-0017). Dropped
  the ~/bin/discord and ~/.claude/skills/Discord provides.files entries (the
  bundle provides them now) and the skill symlink in scripts/postupgrade.sh.

Runtime kept in cortex (the design-doc extraction-fact missed this):
- `cortex network admit` (ADR-0015, O-5) imported role-grant helpers from the
  removed CLI lib (assignRole/resolveRoleId/loadConfig/resolveServerContext).
  The daemon cannot import an external arc bundle, so the narrow runtime slice
  it uses is ported into src/cli/cortex/lib/discord-roles.ts (+ unit tests).
  The grep that "proved" no importers was blind to the relative import path
  (`../../discord/lib/*`); tsc caught it.

Docs:
- docs/agents-md/{architecture,pai-integration}.md updated to say the Discord
  CLI is provided by the bundle; CLAUDE.md regenerated via `arc upgrade compass`.

The live Discord adapter (src/adapters/discord/) STAYS in cortex — only the
CLI tooling moved.

Gate: tsc clean, scoped bun test green (862 + 12 + 177 pass / 0 fail),
check-carveouts PASS, eslint clean, zero live cli/discord references.

Closes #1173
Part of #1171
